receivers: split into smaller files
